implemented cronus

+++ database/migrations/2024_12_16_114836_create_parking_lot_zones_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('parking_lot_zones', function (Blueprint $table) {
             $table->id();
             $table->char('letter', 1);
-            $table->unsignedInteger('price_per_hours');
+            $table->decimal('price_per_hours', 4,2);
             $table->timestamps();
         });
     }
